Set C bindings for unix_rewinddir_job in its own C file
