+++ src/unix/jbuild
@@ -189,6 +189,7 @@
     unix_closedir_job
     unix_valid_dir
     unix_invalidate_dir
+    unix_rewinddir_job
     not_available
   ))
   (install_c_headers (
