+++ src/unix/lwt_unix_windows.h
@@ -122,7 +122,6 @@ CAMLprim value lwt_unix_system_job(value cmdline)
 
 LWT_NOT_AVAILABLE1(unix_readdir_job)
 LWT_NOT_AVAILABLE2(unix_readdir_n_job)
-LWT_NOT_AVAILABLE1(unix_rewinddir_job)
 LWT_NOT_AVAILABLE1(unix_readlink_job)
 LWT_NOT_AVAILABLE3(unix_lockf_job)
 LWT_NOT_AVAILABLE1(unix_gethostname_job)
