+++ src/unix/windows_c/not_available.c
@@ -71,4 +71,5 @@ LWT_NOT_AVAILABLE1(unix_opendir_job)
 LWT_NOT_AVAILABLE1(unix_closedir_job)
 LWT_NOT_AVAILABLE1(unix_valid_dir)
 LWT_NOT_AVAILABLE1(unix_invalidate_dir)
+LWT_NOT_AVAILABLE1(unix_rewinddir_job)
 #endif
